+++ apps/dolly-frontend/src/main/js/src/components/bestillingsveileder/stegVelger/steg/steg1/paneler/Familierelasjoner.js
@@ -1,8 +1,6 @@
 import React from 'react'
-import _get from 'lodash/get'
 import Panel from '~/components/ui/panel/Panel'
 import { Attributt, AttributtKategori } from '../Attributt'
-import Formatters from '~/utils/DataFormatter'
 import { initialSivilstand } from '~/components/fagsystem/pdlf/form/initialValues'
 
 export const FamilierelasjonPanel = ({ stateModifier }) => {
@@ -17,12 +15,6 @@ export const FamilierelasjonPanel = ({ stateModifier }) => {
 		>
 			<AttributtKategori title="Sivilstand">
 				<Attributt attr={sm.attrs.sivilstand} />
-			</AttributtKategori>
-			<AttributtKategori title="Barn">
-				<Attributt attr={sm.attrs.barn} />
-			</AttributtKategori>
-			<AttributtKategori title="Foreldre">
-				<Attributt attr={sm.attrs.foreldre} />
 			</AttributtKategori>
 		</Panel>
 	)
@@ -41,97 +33,4 @@ FamilierelasjonPanel.initialValues = ({ set, del, has, opts }) => ({
 			del('pdldata.person.sivilstand')
 		},
 	},
-	barn: {
-		label: 'Har barn',
-		checked: has('tpsf.relasjoner.barn'),
-		add() {
-			set('tpsf.relasjoner.barn', defaultBarn(opts))
-		},
-		remove() {
-			del('tpsf.relasjoner.barn')
-			!has('tpsf.relasjoner.partnere') && !has('tpsf.relasjoner.foreldre') && del('tpsf.relasjoner')
-		},
-	},
-	foreldre: {
-		label: 'Har foreldre',
-		checked: has('tpsf.relasjoner.foreldre'),
-		add() {
-			set('tpsf.relasjoner.foreldre', defaultForeldre(opts))
-		},
-		remove() {
-			del('tpsf.relasjoner.foreldre')
-			!has('tpsf.relasjoner.partnere') && !has('tpsf.relasjoner.barn') && del('tpsf.relasjoner')
-		},
-	},
 })
-
-const defaultForeldre = (opts) => {
-	const fullForelder = [
-		{
-			identtype: 'FNR',
-			kjonn: '',
-			foreldreType: '',
-			sivilstander: [{ sivilstand: '', sivilstandRegdato: '' }],
-			harFellesAdresse: true,
-			doedsdato: null,
-			alder: Formatters.randomIntInRange(65, 100),
-			spesreg: '',
-			utenFastBopel: false,
-			statsborgerskap: '',
-			statsborgerskapRegdato: '',
-			statsborgerskapTildato: '',
-		},
-	]
-
-	const eksisterendeRelasjoner = _get(opts, 'personFoerLeggTil.tpsf.relasjoner')
-	const eksisterendeForeldre =
-		eksisterendeRelasjoner &&
-		eksisterendeRelasjoner.filter(
-			(relasjon) => relasjon.relasjonTypeNavn === 'MOR' || relasjon.relasjonTypeNavn === 'FAR'
-		)
-	const eksisterendeForelderValues =
-		eksisterendeForeldre &&
-		eksisterendeForeldre.map((forelder) => ({
-			ident: forelder.personRelasjonMed.ident,
-			doedsdato: forelder.personRelasjonMed.doedsdato || null,
-		}))
-
-	return eksisterendeForelderValues && eksisterendeForelderValues.length > 0
-		? eksisterendeForelderValues
-		: fullForelder
-}
-
-const defaultBarn = (opts) => {
-	const fullBarn = [
-		{
-			identtype: 'FNR',
-			kjonn: '',
-			barnType: '',
-			partnerNr: null,
-			borHos: '',
-			erAdoptert: false,
-			alder: Formatters.randomIntInRange(0, 17),
-			doedsdato: null,
-			spesreg: '',
-			utenFastBopel: false,
-			statsborgerskap: '',
-			statsborgerskapRegdato: '',
-			statsborgerskapTildato: '',
-		},
-	]
-
-	const eksisterendeRelasjoner = _get(opts, 'personFoerLeggTil.tpsf.relasjoner')
-	const eksisterendeBarn =
-		eksisterendeRelasjoner &&
-		eksisterendeRelasjoner.filter((relasjon) => relasjon.relasjonTypeNavn === 'FOEDSEL')
-	const eksisterendeBarnValues =
-		eksisterendeBarn &&
-		eksisterendeBarn.map((barn) => ({
-			ident: barn.personRelasjonMed.ident,
-			doedsdato: barn.personRelasjonMed.doedsdato || null,
-		}))
-
-	return eksisterendeBarnValues && eksisterendeBarnValues.length > 0
-		? eksisterendeBarnValues
-		: fullBarn
-}
