+++ apps/dolly-frontend/src/main/js/src/components/bestillingsveileder/stegVelger/steg/steg2/Steg2.js
@@ -3,7 +3,6 @@ import * as Yup from 'yup'
 import { AlertStripeInfo } from 'nav-frontend-alertstriper'
 import { harAvhukedeAttributter } from '~/components/bestillingsveileder/utils'
 import { BestillingsveilederContext } from '~/components/bestillingsveileder/Bestillingsveileder'
-import { TpsfForm } from '~/components/fagsystem/tpsf/form/Form'
 import { KrrstubForm } from '~/components/fagsystem/krrstub/form/Form'
 import { SigrunstubForm } from '~/components/fagsystem/sigrunstub/form/Form'
 import { InntektstubForm } from '~/components/fagsystem/inntektstub/form/Form'
@@ -33,7 +32,6 @@ export const Steg2 = ({ formikBag }) => {
 
 	return (
 		<div>
-			<TpsfForm formikBag={formikBag} />
 			<PdlfForm formikBag={formikBag} />
 			<AaregForm formikBag={formikBag} />
 			<SigrunstubForm formikBag={formikBag} />
@@ -55,7 +53,6 @@ export const Steg2 = ({ formikBag }) => {
 Steg2.label = 'Velg verdier'
 
 Steg2.validation = Yup.object({
-	...TpsfForm.validation,
 	...PdlfForm.validation,
 	...AaregForm.validation,
 	...SigrunstubForm.validation,
