+++ apps/dolly-frontend/src/main/js/src/components/fagsystem/pdlf/form/partials/personinformasjon/Personinformasjon.js
@@ -3,11 +3,10 @@ import Panel from '~/components/ui/panel/Panel'
 import { Vis } from '~/components/bestillingsveileder/VisAttributt'
 import { erForste, panelError } from '~/components/ui/form/formUtils'
 import { Kategori } from '~/components/ui/form/kategori/Kategori'
-import { Diverse } from './partials/Diverse'
-import { Fullmakt } from './partials/fullmakt/Fullmakt'
+import { Fullmakt } from '../fullmakt/Fullmakt'
 import { Sikkerhetstiltak } from '~/components/fagsystem/pdlf/form/partials/sikkerhetstiltak/Sikkerhetstiltak'
 import { BestillingsveilederContext } from '~/components/bestillingsveileder/Bestillingsveileder'
-import { TpsMessagingDiverse } from '~/components/fagsystem/tpsf/form/personinformasjon/partials/tpsmessaging/TpsMessagingDiverse'
+import { TpsMessagingDiverse } from '~/components/fagsystem/tpsmessaging/form/TpsMessagingDiverse'
 import { Telefonnummer } from '~/components/fagsystem/pdlf/form/partials/telefonnummer/Telefonnummer'
 import { Doedsfall } from '~/components/fagsystem/pdlf/form/partials/doedsfall/Doedsfall'
 import { Statsborgerskap } from '~/components/fagsystem/pdlf/form/partials/statsborgerskap/Statsborgerskap'
@@ -19,6 +18,8 @@ import { Kjoenn } from '~/components/fagsystem/pdlf/form/partials/kjoenn/Kjoenn'
 import { Navn } from '~/components/fagsystem/pdlf/form/partials/navn/Navn'
 import { Foedsel } from '~/components/fagsystem/pdlf/form/partials/foedsel/Foedsel'
 import { Vergemaal } from '~/components/fagsystem/pdlf/form/partials/vergemaal/Vergemaal'
+import { UtenlandskBankkonto } from '~/components/fagsystem/tpsmessaging/form/utenlandskbankkonto/UtenlandskBankkonto'
+import { NorskBankkonto } from '~/components/fagsystem/tpsmessaging/form/norskbankkonto/NorskBankkonto'
 
 const nasjonalitetPaths = [
 	'pdldata.person.statsborgerskap',
@@ -27,19 +28,6 @@ const nasjonalitetPaths = [
 ]
 
 const diversePaths = [
-	'tpsf.identtype',
-	'tpsf.harMellomnavn',
-	'tpsf.harNyttNavn',
-	'tpsf.sivilstand',
-	'tpsf.sprakKode',
-	'tpsf.egenAnsattDatoFom',
-	'tpsf.egenAnsattDatoTom',
-	'tpsf.spesreg',
-	'tpsf.utenFastBopel',
-	'tpsf.erForsvunnet',
-	'tpsMessaging.utenlandskBankkonto',
-	'tpsMessaging.norskBankkonto',
-	'tpsMessaging.sikkerhetstiltak',
 	'tpsMessaging.spraakKode',
 	'tpsMessaging.egenAnsattDatoFom',
 	'tpsMessaging.egenAnsattDatoTom',
@@ -51,6 +39,8 @@ const alderPaths = [
 	'pdldata.opprettNyPerson.foedtFoer',
 ]
 
+const utenlandskBankkontoPath = ['tpsMessaging.utenlandskBankkonto']
+const norskBankkontoPath = ['tpsMessaging.norskBankkonto']
 const kjoennPath = ['pdldata.person.kjoenn']
 const navnPath = ['pdldata.person.navn']
 const telefonnummerPath = ['pdldata.person.telefonnummer']
@@ -62,14 +52,7 @@ const foedselPath = ['pdldata.person.foedsel']
 const doedsfallPath = ['pdldata.person.doedsfall']
 const vergemaalPath = ['pdldata.person.vergemaal']
 const fullmaktPath = ['pdldata.person.fullmakt']
-
-const sikkerhetstiltakPaths = [
-	'pdldata.person.sikkerhetstiltak',
-	'tpsf.typeSikkerhetTiltak',
-	'tpsf.beskrSikkerhetTiltak',
-	'tpsf.sikkerhetTiltakDatoFom',
-	'tpsf.sikkerhetTiltakDatoTom',
-]
+const sikkerhetstiltakPath = ['pdldata.person.sikkerhetstiltak']
 
 const panelPaths = [
 	alderPaths,
@@ -85,8 +68,10 @@ const panelPaths = [
 	doedsfallPath,
 	vergemaalPath,
 	fullmaktPath,
-	sikkerhetstiltakPaths,
+	sikkerhetstiltakPath,
 	statsborgerskapPath,
+	utenlandskBankkontoPath,
+	norskBankkontoPath,
 ].flat()
 
 export const Personinformasjon = ({ formikBag }) => {
@@ -98,21 +83,7 @@ export const Personinformasjon = ({ formikBag }) => {
 				heading="Personinformasjon"
 				hasErrors={panelError(formikBag, panelPaths)}
 				iconType={'personinformasjon'}
-				startOpen={() =>
-					erForste(
-						formikBag.values,
-						alderPaths.concat(
-							nasjonalitetPaths,
-							diversePaths,
-							vergemaalPath,
-							fullmaktPath,
-							doedsfallPath,
-							statsborgerskapPath,
-							kjoennPath,
-							navnPath
-						)
-					)
-				}
+				startOpen={() => erForste(formikBag.values, panelPaths)}
 			>
 				{!personFoerLeggTil && (
 					<Kategori title="Alder" vis={alderPaths}>
@@ -151,8 +122,15 @@ export const Personinformasjon = ({ formikBag }) => {
 				</Kategori>
 
 				<Kategori title="Diverse" vis={diversePaths}>
-					<Diverse formikBag={formikBag} />
 					<TpsMessagingDiverse formikBag={formikBag} />
+				</Kategori>
+
+				<Kategori title="Norsk bankkonto" vis={norskBankkontoPath}>
+					<NorskBankkonto formikBag={formikBag} />
+				</Kategori>
+
+				<Kategori title="Utenlandsk bankkonto" vis={utenlandskBankkontoPath}>
+					<UtenlandskBankkonto />
 				</Kategori>
 
 				<Kategori title="Telefonnummer" vis={telefonnummerPath}>
@@ -167,7 +145,7 @@ export const Personinformasjon = ({ formikBag }) => {
 					<Fullmakt formikBag={formikBag} />
 				</Kategori>
 
-				<Kategori title="Sikkerhetstiltak" vis={sikkerhetstiltakPaths}>
+				<Kategori title="Sikkerhetstiltak" vis={sikkerhetstiltakPath}>
 					<Sikkerhetstiltak formikBag={formikBag} />
 				</Kategori>
