--- apps/dolly-frontend/src/main/js/src/components/fagsystem/tpsf/form/personinformasjon/partials/alder/alder.less
@@ -1,18 +1,0 @@
-.alder-component {
-	width: 100%;
-	// display: grid;
-	// grid-template-columns: 1fr 1fr;
-	// grid-column-gap: 20px;
-
-	.inputPanelGruppe__inner {
-		display: grid;
-		grid-template-columns: 1fr 1fr;
-		grid-column-gap: 20px;
-	}
-
-	.alder-felter {
-		margin-top: 12px;
-		display: flex;
-		flex-wrap: wrap;
-	}
-}
