--- apps/dolly-frontend/src/main/js/src/components/fagsystem/tpsf/form/personinformasjon/partials/diskresjonskoder/diskresjonskoder.less
@@ -1,9 +1,0 @@
-.spesreg-component {
-  display: flex;
-  flex-wrap: wrap;
-}
-
-.alertstripe {
-  margin-bottom: 10px;
-  padding-bottom: 10px;
-}
