+++ apps/dolly-frontend/src/main/js/src/components/fagsystem/tpsf/form/validation.js
@@ -1,8 +1,8 @@
 import * as Yup from 'yup'
 import _get from 'lodash/get'
-import { addDays, differenceInWeeks, isAfter, isBefore, isSameDay } from 'date-fns'
+import { addDays, isAfter, isBefore } from 'date-fns'
 import Dataformatter from '~/utils/DataFormatter'
-import { ifPresent, messages, requiredDate, requiredString } from '~/utils/YupValidations'
+import { messages } from '~/utils/YupValidations'
 
 export const sivilstander = Yup.array().of(
 	Yup.object({
@@ -76,303 +76,3 @@ export const sivilstander = Yup.array().of(
 			.required(messages.required),
 	})
 )
-
-const innvandringUtvandringDatoTest = (schema) => {
-	return schema.test(
-		'datoEtterSisteInnUtvandring',
-		`Datoen må være etter siste inn-/utvandring (${''})`,
-		function erEtterSisteDato(dato) {
-			if (isAfter(new Date(dato), new Date())) {
-				return this.createError({
-					message: 'Dato kan ikke være etter dagens dato',
-					path: this.options.path,
-				})
-			}
-
-			const values = this.options.context
-			const personFoerLeggTil = values.personFoerLeggTil
-			if (!dato || !personFoerLeggTil) return true
-
-			const sisteDato = _get(personFoerLeggTil, 'tpsf.innvandretUtvandret[0].flyttedato')
-			if (!sisteDato) return true
-
-			const dateValid = isAfter(new Date(dato), new Date(sisteDato))
-
-			return (
-				dateValid ||
-				this.createError({
-					message: `Dato må være etter siste inn-/utvandring (${Dataformatter.formatDate(
-						sisteDato
-					)})`,
-					path: this.options.path,
-				})
-			)
-		}
-	)
-}
-
-const foedtFoerOgEtterTest = (mainValidation, validerFoedtFoer) => {
-	const errorMsgFoedtFoer =
-		'Født Før dato kan ikke være før Født Etter dato eller etter dagens dato'
-	const errorMsgFoedtEtter =
-		'Født Etter dato kan ikke være etter Født Før dato eller etter dagens dato.'
-
-	return mainValidation.test(
-		'range',
-		validerFoedtFoer ? errorMsgFoedtFoer : errorMsgFoedtEtter,
-		function isWithinTest(val) {
-			if (!val) return true
-
-			const values = this.options.context
-			const path = this.path.substring(0, this.path.lastIndexOf('.'))
-
-			const selectedDato = new Date(new Date(val).toDateString())
-			const foedtEtterValue = _get(values, `${path}.foedtEtter`)
-			const foedtFoerValue = _get(values, `${path}.foedtFoer`)
-
-			const foedtEtterDato = new Date(foedtEtterValue)
-			const foedtFoerDato = new Date(foedtFoerValue)
-
-			const identtype = _get(values, `${path}.identtype`)
-			if (identtype === 'FDAT') {
-				return true
-			}
-
-			if (
-				(!validerFoedtFoer && foedtEtterDato > Date.now()) ||
-				(validerFoedtFoer && foedtFoerDato > Date.now())
-			)
-				return false
-
-			if (foedtEtterDato)
-				if (validerFoedtFoer) {
-					if (foedtEtterValue !== '' && foedtEtterValue !== undefined) {
-						foedtEtterDato.setDate(foedtEtterDato.getDate())
-						if (selectedDato < new Date(foedtEtterDato.toDateString())) return false
-					}
-				} else {
-					if (foedtFoerValue !== '' && foedtFoerValue !== undefined) {
-						foedtFoerDato.setDate(foedtFoerDato.getDate())
-						if (selectedDato > new Date(foedtFoerDato.toDateString())) return false
-					}
-				}
-			return true
-		}
-	)
-}
-
-const requiredHvisIkkeDoed = (path) =>
-	ifPresent(
-		path,
-		Yup.mixed().when('doedsdato', {
-			is: (val) => val === undefined || val === null,
-			then: requiredString,
-		})
-	)
-
-const requiredHvisDoedfoedt = (path) =>
-	ifPresent(
-		path,
-		Yup.mixed().when('identtype', {
-			is: (val) => val === 'FDAT',
-			then: requiredDate,
-		})
-	)
-
-const barn = Yup.array()
-	.of(
-		Yup.object({
-			identtype: Yup.string(),
-			kjonn: Yup.string().nullable(),
-			barnType: ifPresent('$tpsf.relasjoner.barn[0].barnType', requiredString),
-			partnerNr: Yup.number().nullable(),
-			borHos: requiredHvisIkkeDoed('$tpsf.relasjoner.barn[0].borHos'),
-			erAdoptert: Yup.boolean(),
-			alder: Yup.number()
-				.transform((num) => (isNaN(num) ? undefined : num))
-				.min(0, 'Alder må være et positivt tall')
-				.max(119, 'Alder må være under 120'),
-			foedtEtter: foedtFoerOgEtterTest(Yup.date().nullable(), false),
-			foedtFoer: foedtFoerOgEtterTest(Yup.date().nullable(), true),
-			spesreg: Yup.string()
-				.when('utenFastBopel', {
-					is: true,
-					then: Yup.string().test(
-						'is-not-kode6',
-						'Kan ikke være "Kode 6" når "Uten fast bopel" er valgt.',
-						(value) => value !== 'SPSF'
-					),
-				})
-				.nullable(),
-			utenFastBopel: Yup.boolean(),
-			boadresse: Yup.object({
-				kommunenr: Yup.string().nullable(),
-			}),
-			foedselsdato: requiredHvisDoedfoedt('$tpsf.relasjoner.barn[0].foedselsdato'),
-			doedsdato: Yup.date().nullable(),
-		})
-	)
-	.nullable()
-
-const foreldre = Yup.array()
-	.of(
-		Yup.object({
-			identtype: Yup.string(),
-			kjonn: Yup.string().nullable(),
-			foreldreType: ifPresent('$tpsf.relasjoner.foreldre[0].foreldreType', requiredString),
-			alder: Yup.number()
-				.transform((num) => (isNaN(num) ? undefined : num))
-				.min(0, 'Alder må være et positivt tall')
-				.max(119, 'Alder må være under 120'),
-			foedtEtter: foedtFoerOgEtterTest(Yup.date().nullable(), false),
-			foedtFoer: foedtFoerOgEtterTest(Yup.date().nullable(), true),
-			spesreg: Yup.string(),
-			boadresse: Yup.object({
-				kommunenr: Yup.string().nullable(),
-			}),
-			doedsdato: Yup.date().nullable(),
-		})
-	)
-	.nullable()
-
-export const validation = {
-	tpsf: ifPresent(
-		'$tpsf',
-		Yup.object({
-			alder: Yup.number()
-				.min(0, 'Alder må være et positivt tall')
-				.max(119, 'Alder må være under 120')
-				.typeError(messages.required),
-			foedtEtter: foedtFoerOgEtterTest(Yup.date().nullable(), false),
-			foedtFoer: foedtFoerOgEtterTest(Yup.date().nullable(), true),
-			doedsdato: Yup.date().nullable(),
-			kjonn: ifPresent('$tpsf.kjonn', requiredString),
-			statsborgerskap: ifPresent('$tpsf.statsborgerskap', requiredString),
-			statsborgerskapRegdato: Yup.date().nullable(),
-			statsborgerskapTildato: Yup.date().nullable(),
-			innvandretFraLand: ifPresent('$tpsf.innvandretFraLand', requiredString),
-			innvandretFraLandFlyttedato: ifPresent(
-				'$tpsf.innvandretFraLandFlyttedato',
-				innvandringUtvandringDatoTest(Yup.date().nullable())
-			),
-			utvandretTilLand: ifPresent('$tpsf.utvandretTilLand', requiredString),
-			utvandretTilLandFlyttedato: ifPresent(
-				'$tpsf.utvandretTilLandFlyttedato',
-				innvandringUtvandringDatoTest(Yup.date().nullable())
-			),
-			sprakKode: ifPresent('$tpsf.sprakKode', requiredString),
-			egenAnsattDatoFom: ifPresent(
-				'$tpsf.egenAnsattDatoFom',
-				Yup.string().test(
-					'is-before-today',
-					'Dato kan ikke være etter dagens dato',
-					function validDate(dato) {
-						return isBefore(new Date(dato), new Date())
-					}
-				)
-			),
-			egenAnsattDatoTom: ifPresent(
-				'$tpsf.egenAnsattDatoTom',
-				Yup.string().test(
-					'is-after-dato-fom',
-					'Dato må være etter fra-dato',
-					function validDate(dato) {
-						const values = this.options.context
-						return isAfter(new Date(dato), new Date(_get(values, 'tpsf.egenAnsattDatoFom')))
-					}
-				)
-			),
-			spesreg: ifPresent(
-				'$tpsf.spesreg',
-				Yup.string().when('utenFastBopel', {
-					is: true,
-					then: Yup.string().test(
-						'is-not-kode6',
-						'Kan ikke være "Kode 6" når "Uten fast bopel" er valgt.',
-						(value) => value !== 'SPSF'
-					),
-				})
-			).nullable(),
-			typeSikkerhetTiltak: ifPresent('$tpsf.typeSikkerhetTiltak', requiredString),
-			beskrSikkerhetTiltak: ifPresent('$tpsf.beskrSikkerhetTiltak', requiredString),
-			sikkerhetTiltakDatoFom: ifPresent('$tpsf.sikkerhetTiltakDatoFom', requiredDate),
-			sikkerhetTiltakDatoTom: ifPresent(
-				'$tpsf.sikkerhetTiltakDatoTom',
-				Yup.string()
-					.test(
-						'is-after-startdato',
-						'Dato må være lik eller etter startdato, og ikke mer enn 12 uker etter startdato',
-						function validDate(dato) {
-							const values = this.options.context
-							return (
-								(isAfter(new Date(dato), new Date(_get(values, 'tpsf.sikkerhetTiltakDatoFom'))) ||
-									isSameDay(
-										new Date(dato),
-										new Date(_get(values, 'tpsf.sikkerhetTiltakDatoFom'))
-									)) &&
-								differenceInWeeks(
-									new Date(dato),
-									new Date(_get(values, 'tpsf.sikkerhetTiltakDatoFom'))
-								) <= 12
-							)
-						}
-					)
-					.nullable()
-			),
-			relasjoner: Yup.object({
-				barn: ifPresent('$tpsf.relasjoner.barn', barn),
-				foreldre: ifPresent('$tpsf.relasjoner.foreldre', foreldre),
-			}),
-			identtype: ifPresent('$tpsf.identtype', requiredString),
-		})
-	),
-	tpsMessaging: ifPresent(
-		'$tpsMessaging',
-		Yup.object({
-			egenAnsattDatoFom: ifPresent(
-				'$tpsMessaging.egenAnsattDatoFom',
-				Yup.string().test(
-					'is-before-today',
-					'Dato kan ikke være etter dagens dato',
-					function validDate(dato) {
-						return isBefore(new Date(dato), new Date())
-					}
-				)
-			),
-			egenAnsattDatoTom: ifPresent(
-				'$tpsMessaging.egenAnsattDatoTom',
-				Yup.string().test(
-					'is-after-dato-fom',
-					'Dato må være etter fra-dato og senest dagens dato',
-					function validDate(dato) {
-						const values = this.options.context
-						return (
-							isAfter(new Date(dato), new Date(_get(values, 'tpsMessaging.egenAnsattDatoFom'))) &&
-							!isAfter(new Date(dato), new Date())
-						)
-					}
-				)
-			),
-			utenlandskBankkonto: ifPresent(
-				'$tpsMessaging.utenlandskBankkonto',
-				Yup.object().shape({
-					kontonummer: requiredString.nullable(),
-					swift: Yup.string().nullable().optional(),
-					landkode: requiredString.nullable(),
-					iban: Yup.string().nullable().optional(),
-					valuta: requiredString.nullable(),
-					banknavn: Yup.string().nullable().optional(),
-					bankAdresse1: Yup.string().nullable().optional(),
-					bankAdresse2: Yup.string().nullable().optional(),
-					bankAdresse3: Yup.string().nullable().optional(),
-				})
-			),
-			norskBankkonto: ifPresent(
-				'$tpsMessaging.norskBankkonto',
-				Yup.object().shape({
-					kontonummer: requiredString.nullable(),
-				})
-			),
-		})
-	),
-}
