+++ apps/dolly-frontend/src/main/js/src/components/fagsystem/tpsmessaging/form/TpsMessagingDiverse.js
@@ -3,12 +3,7 @@ import { isAfter } from 'date-fns'
 import { PersoninformasjonKodeverk } from '~/config/kodeverk'
 import { FormikDatepicker } from '~/components/ui/form/inputs/datepicker/Datepicker'
 import { FormikSelect } from '~/components/ui/form/inputs/select/Select'
-import { Kategori } from '~/components/ui/form/kategori/Kategori'
-import { UtenlandskBankkonto } from '~/components/fagsystem/tpsf/form/personinformasjon/partials/tpsmessaging/utenlandskbankkonto/UtenlandskBankkonto'
-import { NorskBankkonto } from '~/components/fagsystem/tpsf/form/personinformasjon/partials/tpsmessaging/norskbankkonto/NorskBankkonto'
 import { BestillingsveilederContext } from '~/components/bestillingsveileder/Bestillingsveileder'
-
-const bankkontoPath = ['tpsMessaging.utenlandskBankkonto', 'tpsMessaging.norskBankkonto']
 
 export const TpsMessagingDiverse = ({ formikBag }) => {
 	const { personFoerLeggTil } = useContext(BestillingsveilederContext)
@@ -62,10 +57,6 @@ export const TpsMessagingDiverse = ({ formikBag }) => {
 					visHvisAvhuket
 				/>
 			)}
-			<Kategori title="Bankkonto" vis={bankkontoPath}>
-				<UtenlandskBankkonto />
-				<NorskBankkonto formikBag={formikBag} />
-			</Kategori>
 		</React.Fragment>
 	)
 }
