+++ apps/dolly-frontend/src/main/js/src/components/fagsystem/tpsmessaging/form/norskbankkonto/NorskBankkonto.tsx
@@ -5,11 +5,6 @@ import { FormikSelect } from '~/components/ui/form/inputs/select/Select'
 import { generateValidKontoOptions } from '~/utils/GenererGyldigNorskBankkonto'
 import _get from 'lodash/get'
 import { FormikProps } from 'formik'
-import styled from 'styled-components'
-
-const Title = styled.h4`
-	width: 100%;
-`
 
 export const NorskBankkonto = ({ formikBag }: { formikBag: FormikProps<{}> }) => {
 	const [validKontoOptions, setValidKontoOptions] = useState([])
@@ -25,7 +20,6 @@ export const NorskBankkonto = ({ formikBag }: { formikBag: FormikProps<{}> }) =>
 	return (
 		<Vis attributt={path} formik>
 			<div className="flexbox--flex-wrap">
-				<Title>Norsk bankkonto</Title>
 				<FormikSelect
 					value={kontonummer}
 					placeholder={kontonummer ? kontonummer : 'Velg..'}
