+++ apps/dolly-frontend/src/main/js/src/components/ui/form/formUtils.js
@@ -19,51 +19,34 @@ export const SyntEvent = (name, value) => ({ target: { name, value } })
 
 export const erForste = (values, attributt) => {
 	const rootPaths = [
-		'tpsf.alder',
-		'tpsf.identtype',
-		'tpsf.foedtEtter',
-		'tpsf.foedtFoer',
-		'tpsf.doedsdato',
-		'tpsf.statsborgerskap',
-		'tpsf.innvandretFraLand',
-		'tpsf.utvandretTilLand',
+		'pdldata.opprettNyPerson.alder',
+		'pdldata.person.foedsel',
+		'pdldata.person.doedsfall',
+		'pdldata.person.statsborgerskap',
+		'pdldata.person.innflytting',
+		'pdldata.person.utflytting',
 		'pdldata.person.kjoenn',
 		'pdldata.person.navn',
-		'tpsf.harNyttNavn',
-		'tpsf.sivilstand',
-		'tpsf.sprakKode',
-		'tpsf.egenAnsattDatoFom',
-		'tpsf.egenAnsattDatoTom',
-		'tpsf.spesreg',
-		'tpsf.erForsvunnet',
-		'tpsf.identHistorikk',
-		'tpsf.vergemaal',
-		'tpsf.fullmakt',
-		'tpsMessaging.utenlandskBankkonto',
-		'tpsMessaging.harBankkontonr',
-		'tpsMessaging.sikkerhetstiltak',
+		'pdldata.person.telefonnummer',
+		'pdldata.person.vergemaal',
+		'pdldata.person.fullmakt',
+		'pdldata.person.sikkerhetstiltak',
+		'pdldata.person.tilrettelagtKommunikasjon',
 		'tpsMessaging.spraakKode',
 		'tpsMessaging.egenAnsattDatoFom',
 		'tpsMessaging.egenAnsattDatoTom',
+		'tpsMessaging.utenlandskBankkonto',
+		'tpsMessaging.norskBankkonto',
 		'pdldata.person.bostedsadresse',
 		'pdldata.person.oppholdsadresse',
 		'pdldata.person.kontaktadresse',
 		'pdldata.person.adressebeskyttelse',
-		'tpsf.relasjoner',
+		'pdldata.person.sivilstand',
 		'aareg',
 		'sigrunstub',
 		'inntektstub',
 		'sykemelding',
 		'brregstub',
-		'pdldata.person.telefonnummer',
-		'pdldata.person.tilrettelagtKommunikasjon',
-		'pdldata.person.doedsfall',
-		'pdldata.person.statsborgerskap',
-		'pdldata.person.innflytting',
-		'pdldata.person.utflytting',
-		'pdlforvalter.falskIdentitet',
-		'pdlforvalter.utenlandskIdentifikasjonsnummer',
-		'pdlforvalter.kontaktinformasjonForDoedsbo',
 		'instdata',
 		'krrstub',
 		'arenaforvalter',
